% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conditional_sd.R
\name{compute_conditional_sd.PosteriorBSVARHMSH}
\alias{compute_conditional_sd.PosteriorBSVARHMSH}
\title{Computes posterior draws of structural shock conditional standard deviations}
\usage{
\method{compute_conditional_sd}{PosteriorBSVARHMSH}(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of class 
\code{PosteriorBSVARHMSH} obtained by running the \code{estimate} function.}
}
\value{
An object of class \code{PosteriorSigma}, that is, an \code{NxTxS} 
array with attribute \code{PosteriorSigma} containing \code{S} draws of the 
structural shock conditional standard deviations.
}
\description{
Each of the draws from the posterior estimation of models is 
transformed into a draw from the posterior distribution of the structural 
shock conditional standard deviations.
}
\examples{
specification  = specify_bsvar_hmsh$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)
csd     = compute_conditional_sd(posterior)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_hmsh$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_conditional_sd() -> csd
  
}
\seealso{
\code{\link{estimate}}, \code{\link{normalise_posterior}}, \code{\link{summary}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
