% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conditional_sd.R
\name{compute_conditional_sd}
\alias{compute_conditional_sd}
\title{Computes posterior draws of structural shock conditional standard deviations}
\usage{
compute_conditional_sd(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of either of the classes: 
PosteriorBSVAR, PosteriorBSVAR-MSH, PosteriorBSVAR-MIX, or PosteriorBSVAR-SV
obtained by running one of the \code{estimate_bsvar_*} functions.}
}
\value{
An object of class PosteriorSigma, that is, an \code{NxTxS} array with attribute PosteriorSigma 
containing \code{S} draws of the structural shock conditional standard deviations.
}
\description{
Each of the draws from the posterior estimation of a model is transformed into
a draw from the posterior distribution of the structural shock conditional standard deviations.
}
\examples{
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)

# run the burn-in
burn_in        = estimate_bsvar(specification, 10)

# estimate the model
posterior      = estimate_bsvar(burn_in$get_last_draw(), 50)

# compute impulse responses 5 years ahead
sigma          = compute_conditional_sd(posterior)

}
\seealso{
\code{\link{estimate_bsvar}}, \code{\link{estimate_bsvar_msh}}, \code{\link{estimate_bsvar_sv}}, \code{\link{estimate_bsvar_mix}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
