% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_historical_decompositions.R
\name{compute_historical_decompositions.PosteriorBSVARSV}
\alias{compute_historical_decompositions.PosteriorBSVARSV}
\title{Computes posterior draws of historical decompositions}
\usage{
\method{compute_historical_decompositions}{PosteriorBSVARSV}(posterior, show_progress = TRUE)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of class 
\code{PosteriorBSVARSV} obtained by running the \code{estimate} function.}

\item{show_progress}{a logical value, if \code{TRUE} the estimation progress bar is visible}
}
\value{
An object of class \code{PosteriorHD}, that is, an \code{NxNxTxS} array 
with attribute \code{PosteriorHD} containing \code{S} draws of the historical 
decompositions.
}
\description{
Each of the draws from the posterior estimation of models from
packages \pkg{bsvars} or \pkg{bsvarSIGNs} is transformed into
a draw from the posterior distribution of the historical decompositions. 
IMPORTANT! The historical decompositions are interpreted correctly for 
covariance stationary data. Application to unit-root non-stationary data might
result in non-interpretable outcomes.
}
\examples{
specification  = specify_bsvar_sv$new(us_fiscal_lsuw, p = 1)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# compute historical decompositions
hd             = compute_historical_decompositions(posterior)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_sv$new(p = 1) |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_historical_decompositions() -> hds
  
}
\references{
Kilian, L., & Lütkepohl, H. (2017). Structural VAR Tools, Chapter 4, In: Structural vector autoregressive analysis. Cambridge University Press.
}
\seealso{
\code{\link{estimate}}, \code{\link{normalise}}, \code{\link{summary}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
