% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_structural_shocks.R
\name{compute_structural_shocks}
\alias{compute_structural_shocks}
\title{Computes posterior draws of structural shocks}
\usage{
compute_structural_shocks(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome obtained by running the \code{estimate} function. 
The interpretation depends on the normalisation of the shocks
using function \code{normalise()}. Verify if the default settings are appropriate.}
}
\value{
An object of class PosteriorShocks, that is, an \code{NxTxS} array with attribute PosteriorShocks 
containing \code{S} draws of the structural shocks.
}
\description{
Each of the draws from the posterior estimation of models from
packages \pkg{bsvars} or \pkg{bsvarSIGNs} is transformed into
a draw from the posterior distribution of the structural shocks.
}
\examples{
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 1)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# compute structural shocks
shocks         = compute_structural_shocks(posterior)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar$new(p = 1) |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_structural_shocks() -> ss

}
\seealso{
\code{\link{estimate}}, \code{\link{normalise}}, \code{\link{summary}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
