% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logSDDR_homoskedasticity}
\alias{logSDDR_homoskedasticity}
\title{The log of Bayes factor for the homoskedasticity hypothesis for each structural shock}
\usage{
logSDDR_homoskedasticity(posterior, prior, Y, X, sample_s_ = TRUE)
}
\arguments{
\item{posterior}{the \code{posterior} element of the estimation outcome from function \code{bsvar_sv}}

\item{prior}{A list specifying the prior distribution. See argument \code{prior} of function \code{bsvar_sv}}

\item{Y}{an \code{NxT} matrix, the matrix containing \code{T} observations on \code{N} dependent time series variables}

\item{X}{a \code{KxT} matrix, the matrix containing \code{T} observations on \code{K = N*p+d} regressors including \code{p} lags of dependent variables and \code{d} deterministic terms}

\item{a}{logical value set to the same value as the corresponding argument of function \code{bsvar_sv}}
}
\value{
A list of two components:

\code{logSDDR} an \code{N}-vector with values of the logarithm of the Bayes factors

\code{components} a list of three components
\describe{
  \item{log_denominator}{an \code{N}-vector with values of the logarithm of the Bayes factors' denominator}
  \item{log_numerator}{an \code{N}-vector with values of the logarithm of the Bayes factors' numerator}
  \item{log_numerator_s}{an \code{NxS} matrix of the log-full conditional posterior density ordinates computed to estimate the numerator}
}
}
\description{
Computes the logarithm of Bayes factor for the homoskedasticity hypothesis 
for each of the structural shocks via Savage-Dickey Density Ration (SDDR).
The hypothesis is represented by restriction:
\Sexpr[results=rd, stage=build]{katex::math_to_rd("w_n = 0")}
The Bayes factor for this hypothesis can be computed from using the SDDR:
\Sexpr[results=rd, stage=build]{katex::math_to_rd("log p(w_n = 0 | data) - log p(w_n = 0)")}
Therefore, negative values of the difference above are the evidence against 
homoskedasticity of the structural shock. The estimation of both values requires 
numerical integration.
}
\references{
Lütkepohl, H., Shang, F., Uzeda, L., and Woźniak, T. (2022) Partial Identification of Heteroskedastic Structural VARs: Theory and Bayesian Inference.
}
\seealso{
[bsvar_sv()]
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
