% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise.PosteriorBSVAREXH}
\alias{normalise.PosteriorBSVAREXH}
\title{Waggoner & Zha (2003) row signs normalisation of the posterior draws 
for the structural matrix \eqn{B}}
\usage{
\method{normalise}{PosteriorBSVAREXH}(posterior, B_benchmark = NULL)
}
\arguments{
\item{posterior}{posterior estimation outcome of class \code{PosteriorBSVAREXH} 
generated using the \code{estimate()} function, amongst other draws, 
the \code{S} draws from the posterior distribution of the \code{NxN} 
structural matrix of contemporaneous relationships \eqn{B}. These draws are 
to be normalised with respect to the matrix \code{B_benchmark}.}

\item{B_benchmark}{the benchmark \code{NxN} structural matrix specified by 
the user to have the desired row signs}
}
\value{
An object of the same class as that provided as the input argument 
\code{posterior} containing the posterior draws including the draws of the 
normalised structural matrix.
}
\description{
Normalises the sign of rows of matrix \eqn{B} MCMC draws, 
 relative to matrix \code{B_benchmark}, provided as the second argument. The implemented
 procedure proposed by Waggoner, Zha (2003) normalises the MCMC output in an
 optimal way leading to the unimodal posterior. Only normalised MCMC output is 
 suitable for the computations of the posterior characteristics of the \eqn{B}
 matrix elements and their functions such as the impulse response functions and other 
 economically interpretable values.
}
\examples{
specification  = specify_bsvar_exh$new(us_fiscal_lsuw)    # specify the model
burn_in        = estimate(specification, 5)               # run the burn-in
posterior      = estimate(burn_in, 5)                     # estimate the model

# normalise the posterior
BB             = posterior$last_draw$starting_values$B    # get the last draw of B
B_benchmark          = diag((-1) * sign(diag(BB))) \%*\% BB       # set negative diagonal elements
posterior      = normalise(posterior, B_benchmark)              # draws in posterior are normalised

}
\references{
Waggoner, D.F., and Zha, T., (2003) Likelihood Preserving Normalization in Multiple Equation Models. 
\emph{Journal of Econometrics}, \bold{114}(2), 329--47, \doi{10.1016/S0304-4076(03)00087-3}.
}
\seealso{
\code{\link{estimate}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
