% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PosteriorFitted}
\alias{plot.PosteriorFitted}
\title{Plots fitted values of dependent variables}
\usage{
\method{plot}{PosteriorFitted}(
  x,
  probability = 0.9,
  col = "#ff69b4",
  main,
  xlab,
  mar.multi = c(1, 4.6, 0, 2.1),
  oma.multi = c(6, 0, 5, 0),
  ...
)
}
\arguments{
\item{x}{an object of class PosteriorFitted obtained using the
\code{compute_fitted_values()} function containing posterior draws of 
fitted values of dependent variables.}

\item{probability}{a parameter determining the interval to be plotted. The 
interval stretches from the \code{0.5 * (1 - probability)} to 
\code{1 - 0.5 * (1 - probability)} percentile of the posterior distribution.}

\item{col}{a colour of the plot line and the ribbon}

\item{main}{an alternative main title for the plot}

\item{xlab}{an alternative x-axis label for the plot}

\item{mar.multi}{the default \code{mar} argument setting in \code{graphics::par}. Modify with care!}

\item{oma.multi}{the default \code{oma} argument setting in \code{graphics::par}. Modify with care!}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Plots of fitted values of dependent variables including their 
median and percentiles.
}
\examples{
specification  = specify_bsvar$new(us_fiscal_lsuw)    # specify model
burn_in        = estimate(specification, 5)          # run the burn-in
posterior      = estimate(burn_in, 5)                # estimate the model

# compute fitted values
fitted         = compute_fitted_values(posterior)
plot(fitted)                                          # plot fitted values

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_fitted_values() |>
  plot()

}
\seealso{
\code{\link{compute_fitted_values}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
