% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PosteriorHD}
\alias{plot.PosteriorHD}
\title{Plots historical decompositions}
\usage{
\method{plot}{PosteriorHD}(
  x,
  shock_names,
  cols,
  main,
  xlab,
  mar.multi = c(1, 4.6, 0, 4.6),
  oma.multi = c(6, 0, 5, 0),
  ...
)
}
\arguments{
\item{x}{an object of class PosteriorHD obtained using the
\code{compute_historical_decompositions()} function containing posterior draws of 
historical decompositions.}

\item{shock_names}{a vector of length \code{N} containing names of the structural shocks.}

\item{cols}{an \code{N}-vector with colours of the plot}

\item{main}{an alternative main title for the plot}

\item{xlab}{an alternative x-axis label for the plot}

\item{mar.multi}{the default \code{mar} argument setting in \code{graphics::par}. Modify with care!}

\item{oma.multi}{the default \code{oma} argument setting in \code{graphics::par}. Modify with care!}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Plots of the posterior means of the historical decompositions.
}
\examples{
specification  = specify_bsvar$new(us_fiscal_lsuw)    # specify model
burn_in        = estimate(specification, 5)           # run the burn-in
posterior      = estimate(burn_in, 5)                 # estimate the model

# compute historical decompositions
fevd           = compute_historical_decompositions(posterior)
plot(fevd)                                            

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_historical_decompositions() |>
  plot()

}
\seealso{
\code{\link{compute_historical_decompositions}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
