% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvar_mix.R
\name{specify_bsvar_mix}
\alias{specify_bsvar_mix}
\title{R6 Class representing the specification of the BSVAR model with a zero-mean mixture of normals model for structural shocks.}
\description{
The class BSVARMIX presents complete specification for the BSVAR model with a zero-mean mixture of normals model for structural shocks.
}
\examples{
data(us_fiscal_lsuw)
spec = specify_bsvar_mix$new(
   data = us_fiscal_lsuw,
   p = 4,
   M = 2
)

}
\seealso{
\code{\link{estimate}}, \code{\link{specify_posterior_bsvar_mix}}
}
\section{Super class}{
\code{BSVARMSH} -> \code{BSVARMIX}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{p}}{a non-negative integer specifying the autoregressive lag order of the model.}

    \item{\code{identification}}{an object IdentificationBSVARs with the identifying restrictions.}

    \item{\code{prior}}{an object PriorBSVARMIX with the prior specification.}

    \item{\code{data_matrices}}{an object DataMatricesBSVAR with the data matrices.}

    \item{\code{starting_values}}{an object StartingValuesBSVARMIX with the starting values.}

    \item{\code{finiteM}}{a logical value - if true a finite mixture model is estimated. Otherwise, a sparse mixture model is estimated in which \code{M=20} and the number of visited states is estimated.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BSVARMIX-initialize}{\code{BSVARMIX$new()}}
    \item \href{#method-BSVARMIX-clone}{\code{BSVARMIX$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><code>BSVARMSH$get_data_matrices()</code></li>
  <li><code>BSVARMSH$get_identification()</code></li>
  <li><code>BSVARMSH$get_normal()</code></li>
  <li><code>BSVARMSH$get_prior()</code></li>
  <li><code>BSVARMSH$get_starting_values()</code></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARMIX-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARMIX-initialize}{}}}
\subsection{\code{BSVARMIX$new()}}{
  Create a new specification of the BSVAR model with a zero-mean mixture of normals model for structural shocks, BSVARMIX.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BSVARMIX$new(
  data,
  p = 1L,
  M = 2L,
  B,
  A,
  distribution = c("norm", "t"),
  exogenous = NULL,
  stationary = rep(FALSE, ncol(data)),
  finiteM = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{a \code{(T+p)xN} matrix with time series data.}
      \item{\code{p}}{a positive integer providing model's autoregressive lag order.}
      \item{\code{M}}{an integer greater than 1 - the number of components of the mixture of normals.}
      \item{\code{B}}{a logical \code{NxN} matrix containing value \code{TRUE} for the elements of the structural matrix \eqn{B} to be estimated and value \code{FALSE} for exclusion restrictions to be set to zero.}
      \item{\code{A}}{a logical \code{NxK} matrix containing value \code{TRUE} for the elements of 
the autoregressive matrix \eqn{A} to be estimated and value \code{FALSE} for exclusion restrictions 
to be set to zero.}
      \item{\code{distribution}}{a character string specifying the conditional distribution 
of structural shocks. Value \code{"norm"} sets it to the normal distribution, 
while value \code{"t"} sets the Student-t distribution.}
      \item{\code{exogenous}}{a \code{(T+p)xd} matrix of exogenous variables.}
      \item{\code{stationary}}{an \code{N} logical vector - its element set to \code{FALSE} sets 
the prior mean for the autoregressive parameters of the \code{N}th equation to the random walk process, 
otherwise to white noise.}
      \item{\code{finiteM}}{a logical value - if true a finite mixture model is estimated. Otherwise, a sparse mixture model is estimated in which \code{M=20} and the number of visited states is estimated.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new complete specification for the bsvar model with a zero-mean mixture of normals model for structural shocks, BSVARMIX.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARMIX-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARMIX-clone}{}}}
\subsection{\code{BSVARMIX$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BSVARMIX$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
