% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvar_exh.R
\name{specify_posterior_bsvar_exh}
\alias{specify_posterior_bsvar_exh}
\title{R6 Class Representing PosteriorBSVAREXH}
\description{
The class PosteriorBSVAREXH contains posterior output and the specification including 
the last MCMC draw for the bsvar model with exogenous heteroskedastic regime changes. 
Note that due to the thinning of the MCMC output the starting value in element \code{last_draw}
might not be equal to the last draw provided in element \code{posterior}.
}
\examples{
# This is a function that is used within estimate()
spec   = specify_bsvar_exh$new(us_fiscal_lsuw)
post   = estimate(spec, 10, thin = 1)
class(post)


## ------------------------------------------------
## Method `PosteriorBSVAREXH$get_posterior()`
## ------------------------------------------------

spec   = specify_bsvar_exh$new(us_fiscal_lsuw)
post   = estimate(spec, 10, thin = 1)
post$get_posterior()


## ------------------------------------------------
## Method `PosteriorBSVAREXH$get_last_draw()`
## ------------------------------------------------

# specify the model
spec   = specify_bsvar_exh$new(us_fiscal_lsuw)

# run the burn-in
burn   = estimate(spec, 10, thin = 2)

# estimate the model
post   = estimate(burn, 10, thin = 2)


## ------------------------------------------------
## Method `PosteriorBSVAREXH$is_normalised()`
## ------------------------------------------------

# specify the model
spec   = specify_bsvar_exh$new(us_fiscal_lsuw)

# estimate the model
post   = estimate(spec, 10, thin = 1)

# check normalisation status beforehand
post$is_normalised()

# normalise the posterior
BB            = post$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) \%*\% BB    # set negative diagonal elements
post          = normalise(post, B_hat)                # draws in posterior are normalised

# check normalisation status afterwards
post$is_normalised()


## ------------------------------------------------
## Method `PosteriorBSVAREXH$set_normalised()`
## ------------------------------------------------

# This is an internal function that is run while executing normalise()
# Observe its working by analysing the workflow:

spec   = specify_bsvar_exh$new(us_fiscal_lsuw)
post   = estimate(spec, 10, thin = 1)

# check normalisation status beforehand
post$is_normalised()

# normalise the posterior
BB            = post$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) \%*\% BB           # set positive diagonal elements
post          = normalise(post, B_hat)                # draws in posterior are normalised

# check normalisation status afterwards
post$is_normalised()

}
\seealso{
\code{\link{estimate}}, \code{\link{specify_bsvar_exh}}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{last_draw}}{an object of class BSVAREXH with the last draw of the current 
MCMC run as the starting value to be passed to the continuation of the MCMC 
estimation using \code{estimate()}.}

    \item{\code{posterior}}{a list containing Bayesian estimation output.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PosteriorBSVAREXH-initialize}{\code{PosteriorBSVAREXH$new()}}
    \item \href{#method-PosteriorBSVAREXH-get_posterior}{\code{PosteriorBSVAREXH$get_posterior()}}
    \item \href{#method-PosteriorBSVAREXH-get_last_draw}{\code{PosteriorBSVAREXH$get_last_draw()}}
    \item \href{#method-PosteriorBSVAREXH-is_normalised}{\code{PosteriorBSVAREXH$is_normalised()}}
    \item \href{#method-PosteriorBSVAREXH-set_normalised}{\code{PosteriorBSVAREXH$set_normalised()}}
    \item \href{#method-PosteriorBSVAREXH-clone}{\code{PosteriorBSVAREXH$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBSVAREXH-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBSVAREXH-initialize}{}}}
\subsection{\code{PosteriorBSVAREXH$new()}}{
  Create a new posterior output PosteriorBSVAREXH
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBSVAREXH$new(specification_bsvar, posterior_bsvar)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{specification_bsvar}}{an object of class BSVAREXH with the last draw 
of the current MCMC run as the starting value.}
      \item{\code{posterior_bsvar}}{a list containing Bayesian estimation output.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A posterior output PosteriorBSVAREXH
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBSVAREXH-get_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBSVAREXH-get_posterior}{}}}
\subsection{\code{PosteriorBSVAREXH$get_posterior()}}{
  Returns a list containing Bayesian estimation output.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBSVAREXH$get_posterior()}
    \if{html}{\out{</div>}}
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{spec   = specify_bsvar_exh$new(us_fiscal_lsuw)
post   = estimate(spec, 10, thin = 1)
post$get_posterior()
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBSVAREXH-get_last_draw"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBSVAREXH-get_last_draw}{}}}
\subsection{\code{PosteriorBSVAREXH$get_last_draw()}}{
  Returns an object of class BSVAREXH with the last draw of the current MCMC 
run as the starting value to be passed to the continuation of the MCMC 
estimation using \code{estimate()}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBSVAREXH$get_last_draw()}
    \if{html}{\out{</div>}}
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{# specify the model
spec   = specify_bsvar_exh$new(us_fiscal_lsuw)

# run the burn-in
burn   = estimate(spec, 10, thin = 2)

# estimate the model
post   = estimate(burn, 10, thin = 2)
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBSVAREXH-is_normalised"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBSVAREXH-is_normalised}{}}}
\subsection{\code{PosteriorBSVAREXH$is_normalised()}}{
  Returns \code{TRUE} if the posterior has been normalised using \code{normalise()} and \code{FALSE} otherwise.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBSVAREXH$is_normalised()}
    \if{html}{\out{</div>}}
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{# specify the model
spec   = specify_bsvar_exh$new(us_fiscal_lsuw)

# estimate the model
post   = estimate(spec, 10, thin = 1)

# check normalisation status beforehand
post$is_normalised()

# normalise the posterior
BB            = post$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag((-1) * sign(diag(BB))) \%*\% BB    # set negative diagonal elements
post          = normalise(post, B_hat)                # draws in posterior are normalised

# check normalisation status afterwards
post$is_normalised()
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBSVAREXH-set_normalised"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBSVAREXH-set_normalised}{}}}
\subsection{\code{PosteriorBSVAREXH$set_normalised()}}{
  Sets the private indicator \code{normalised} to TRUE.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBSVAREXH$set_normalised(value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{(optional) a logical value to be passed to indicator \code{normalised}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{# This is an internal function that is run while executing normalise()
# Observe its working by analysing the workflow:

spec   = specify_bsvar_exh$new(us_fiscal_lsuw)
post   = estimate(spec, 10, thin = 1)

# check normalisation status beforehand
post$is_normalised()

# normalise the posterior
BB            = post$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) \%*\% BB           # set positive diagonal elements
post          = normalise(post, B_hat)                # draws in posterior are normalised

# check normalisation status afterwards
post$is_normalised()
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PosteriorBSVAREXH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PosteriorBSVAREXH-clone}{}}}
\subsection{\code{PosteriorBSVAREXH$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PosteriorBSVAREXH$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
