% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvar.R
\name{specify_starting_values_bsvar}
\alias{specify_starting_values_bsvar}
\title{R6 Class Representing StartingValuesBSVAR}
\description{
The class StartingValuesBSVAR presents starting values for the homoskedastic bsvar model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{A}}{an \code{NxK} matrix of starting values for the parameter \eqn{A}.}

\item{\code{B}}{an \code{NxN} matrix of starting values for the parameter \eqn{B}.}

\item{\code{hyper}}{a \code{5}-vector of starting values for the shrinkage hyper-parameters of the hierarchical prior distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StartingValuesBSVAR-new}{\code{specify_starting_values_bsvar$new()}}
\item \href{#method-StartingValuesBSVAR-get_starting_values_bsvar}{\code{specify_starting_values_bsvar$get_starting_values_bsvar()}}
\item \href{#method-StartingValuesBSVAR-clone}{\code{specify_starting_values_bsvar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVAR-new"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVAR-new}{}}}
\subsection{Method \code{new()}}{
Create new starting values StartingValuesBSVAR.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_starting_values_bsvar$new(N, p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{p}}{a positive integer - the autoregressive lag order of the SVAR model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Starting values StartingValuesBSVAR.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVAR-get_starting_values_bsvar"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVAR-get_starting_values_bsvar}{}}}
\subsection{Method \code{get_starting_values_bsvar()}}{
Returns the elements of the starting values StartingValuesBSVAR as a \code{list}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_starting_values_bsvar$get_starting_values_bsvar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVAR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVAR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_starting_values_bsvar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
