% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvar_exh.R
\name{specify_starting_values_bsvar_exh}
\alias{specify_starting_values_bsvar_exh}
\title{R6 Class Representing StartingValuesBSVAREXH}
\description{
The class StartingValuesBSVAREXH presents starting values for the bsvar model 
with exogenous regime change Heteroskedasticity.
}
\examples{
# starting values for a bsvar model for a 3-variable system
A = matrix(TRUE, 3, 4)
B = matrix(TRUE, 3, 3)
sv = specify_starting_values_bsvar_exh$new(A = A, B = B, N = 3, p = 1, T = 100)


## ------------------------------------------------
## Method `StartingValuesBSVAREXH$get_starting_values()`
## ------------------------------------------------

# starting values for a homoskedastic bsvar with 1 lag for a 3-variable system
A = matrix(TRUE, 3, 4)
B = matrix(TRUE, 3, 3)
sv = specify_starting_values_bsvar_exh$new(A = A, B = B, N = 3, p = 1, T = 100)
sv$get_starting_values()   # show starting values as list


## ------------------------------------------------
## Method `StartingValuesBSVAREXH$set_starting_values()`
## ------------------------------------------------

# starting values for a bsvar model with 1 lag for a 3-variable system
A = matrix(TRUE, 3, 4)
B = matrix(TRUE, 3, 3)
sv = specify_starting_values_bsvar_exh$new(A = A, B = B, N = 3, p = 1, T = 100)

# Modify the starting values by:
sv_list = sv$get_starting_values()   # getting them as list
sv_list$A <- matrix(rnorm(12), 3, 4) # modifying the entry
sv$set_starting_values(sv_list)      # providing to the class object

}
\section{Super class}{
\code{StartingValuesBSVAR} -> \code{StartingValuesBSVAREXH}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{A}}{an \code{NxK} matrix of starting values for the parameter \eqn{A}.}

    \item{\code{B}}{an \code{NxN} matrix of starting values for the parameter \eqn{B}.}

    \item{\code{hyper}}{a \code{(2*N+1)x2} matrix of starting values for the shrinkage hyper-parameters of the 
hierarchical prior distribution.}

    \item{\code{sigma2}}{an \code{NxM} matrix of starting values for the regime-specific 
variances of the structural shocks. Its elements sum to value \code{M} over the rows.}

    \item{\code{xi}}{a \code{1xT} matrix of starting values for the Markov process
indicator with integer values from \code{0} to \code{M-1}.}

    \item{\code{lambda}}{a \code{NxT} matrix of starting values for latent variables.}

    \item{\code{df}}{an \code{Nx1} vector of positive numbers with starting values 
for the equation-specific degrees of freedom parameters of the Student-t 
conditional distribution of structural shocks.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-StartingValuesBSVAREXH-initialize}{\code{StartingValuesBSVAREXH$new()}}
    \item \href{#method-StartingValuesBSVAREXH-get_starting_values}{\code{StartingValuesBSVAREXH$get_starting_values()}}
    \item \href{#method-StartingValuesBSVAREXH-set_starting_values}{\code{StartingValuesBSVAREXH$set_starting_values()}}
    \item \href{#method-StartingValuesBSVAREXH-clone}{\code{StartingValuesBSVAREXH$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVAREXH-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVAREXH-initialize}{}}}
\subsection{\code{StartingValuesBSVAREXH$new()}}{
  Create new starting values StartingValuesBSVAREXH.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StartingValuesBSVAREXH$new(A, B, N, p, T, d = 0, variance_regimes = rep(1, T))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{A}}{a logical \code{NxK} matrix containing value \code{TRUE} for the elements of 
the autoregressive matrix \eqn{A} to be estimated and value \code{FALSE} for exclusion restrictions 
to be set to zero.}
      \item{\code{B}}{a logical \code{NxN} matrix containing value \code{TRUE} for the elements of 
the structural matrix \eqn{B} to be estimated and value \code{FALSE} for exclusion restrictions 
to be set to zero.}
      \item{\code{N}}{a positive integer - the number of dependent variables in the model.}
      \item{\code{p}}{a positive integer - the autoregressive lag order of the SVAR model.}
      \item{\code{T}}{a positive integer - the the time series dimension of the dependent 
variable matrix \eqn{Y}.}
      \item{\code{d}}{a positive integer - the number of \code{exogenous} variables in the model.}
      \item{\code{variance_regimes}}{a \code{T}-vector with exogenous regime indicators that are integer 
numbers associating the time observation with heteroskedastic regime.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Starting values StartingValuesBSVAREXH.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVAREXH-get_starting_values"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVAREXH-get_starting_values}{}}}
\subsection{\code{StartingValuesBSVAREXH$get_starting_values()}}{
  Returns the elements of the starting values StartingValuesBSVAR-MS as a \code{list}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StartingValuesBSVAREXH$get_starting_values()}
    \if{html}{\out{</div>}}
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{# starting values for a homoskedastic bsvar with 1 lag for a 3-variable system
A = matrix(TRUE, 3, 4)
B = matrix(TRUE, 3, 3)
sv = specify_starting_values_bsvar_exh$new(A = A, B = B, N = 3, p = 1, T = 100)
sv$get_starting_values()   # show starting values as list
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVAREXH-set_starting_values"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVAREXH-set_starting_values}{}}}
\subsection{\code{StartingValuesBSVAREXH$set_starting_values()}}{
  Returns the elements of the starting values StartingValuesBSVAREXH as a \code{list}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StartingValuesBSVAREXH$set_starting_values(last_draw)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{last_draw}}{a list containing the last draw.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    An object of class StartingValuesBSVAREXH including the last draw 
of the current MCMC as the starting value to be passed to the continuation 
of the MCMC estimation using \code{estimate()}.
  }
  \subsection{Examples}{
    \if{html}{\out{<div class="r example copy">}}
    \preformatted{# starting values for a bsvar model with 1 lag for a 3-variable system
A = matrix(TRUE, 3, 4)
B = matrix(TRUE, 3, 3)
sv = specify_starting_values_bsvar_exh$new(A = A, B = B, N = 3, p = 1, T = 100)

# Modify the starting values by:
sv_list = sv$get_starting_values()   # getting them as list
sv_list$A <- matrix(rnorm(12), 3, 4) # modifying the entry
sv$set_starting_values(sv_list)      # providing to the class object
}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVAREXH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVAREXH-clone}{}}}
\subsection{\code{StartingValuesBSVAREXH$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{StartingValuesBSVAREXH$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
