% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Forecasts}
\alias{summary.Forecasts}
\title{Provides posterior summary of Forecasts}
\usage{
\method{summary}{Forecasts}(object, ...)
}
\arguments{
\item{object}{an object of class Forecasts obtained using the
\code{forecast()} function containing draws the predictive density.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 
5 and 95 percentiles of the forecasts for each of the variables and forecast 
horizons.
}
\description{
Provides posterior summary of the forecasts including their 
mean, standard deviations, as well as 5 and 95 percentiles.
}
\examples{
specification  = specify_bsvar$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# forecast
fore           = forecast(posterior, horizon = 2)
fore_summary   = summary(fore)
fore_summary$variable1

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  forecast(horizon = 2) |>
  summary() -> fore_summary
fore_summary$variable1

}
\seealso{
\code{\link{forecast}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
