% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorBSVARMSH}
\alias{summary.PosteriorBSVARMSH}
\title{Provides posterior summary of heteroskedastic Structural VAR estimation}
\usage{
\method{summary}{PosteriorBSVARMSH}(object, ...)
}
\arguments{
\item{object}{an object of class PosteriorBSVARMSH obtained using the
\code{estimate()} function applied to heteroskedastic Bayesian Structural VAR
model specification set by function \code{specify_bsvar_msh$new()} containing 
draws from the  posterior distribution of the parameters.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 5 and 95 
percentiles of the parameters: the structural matrix \eqn{B}, autoregressive 
parameters \eqn{A}, and hyper-parameters.
}
\description{
Provides posterior mean, standard deviations, as well as 5 and 95 
percentiles of the parameters: the structural matrix \eqn{B}, autoregressive 
parameters \eqn{A}, and hyper parameters.
}
\examples{
specification  = specify_bsvar_msh$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)
summ           = summary(posterior)
summ

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_msh$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  summary() -> summ
summ

}
\seealso{
\code{\link{estimate}}, \code{\link{specify_bsvar_msh}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
