% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorFEVD}
\alias{summary.PosteriorFEVD}
\title{Provides posterior summary of forecast error variance decompositions}
\usage{
\method{summary}{PosteriorFEVD}(object, ...)
}
\arguments{
\item{object}{an object of class PosteriorFEVD obtained using the
\code{compute_variance_decompositions()} function containing draws from the 
posterior distribution of the forecast error variance decompositions.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean of the forecast error variance 
decompositions of each variable at all horizons.
}
\description{
Provides posterior means of the forecast error variance 
decompositions of each variable at all horizons.
}
\examples{
specification  = specify_bsvar$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# compute forecast error variance decompositions
fevd           = compute_variance_decompositions(posterior, horizon = 4)
fevd_summary   = summary(fevd)
fevd_summary$variable1

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_variance_decompositions(horizon = 4) |>
  summary() -> fevd_summary
fevd_summary$variable1

}
\seealso{
\code{\link{compute_variance_decompositions}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
