% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorRegimePr}
\alias{summary.PosteriorRegimePr}
\title{Provides posterior summary of regimes or regime probabilities}
\usage{
\method{summary}{PosteriorRegimePr}(object, ...)
}
\arguments{
\item{object}{an object of class PosteriorRegimePr obtained using the
\code{compute_regime_probabilities()} function containing posterior draws of 
regime allocations.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean and standard deviations of the
realized regime indices or regime probabilities.
}
\description{
Provides posterior summary of realized regime indices or regime
probabilities, including their mean and standard deviations.
}
\examples{
specification  = specify_bsvar_msh$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# compute regime probabilities
rp             = compute_regime_probabilities(posterior)
rp_summary     = summary(rp)
head(rp_summary$MarkovProcess1$regime) # browse the results

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_msh$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_regime_probabilities() |>
  summary() -> rp_summary
head(rp_summary$MarkovProcess1$regime) # browse the results

}
\seealso{
\code{\link{compute_regime_probabilities}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
