% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_identification.PosteriorBSVARMSH}
\alias{verify_identification.PosteriorBSVARMSH}
\title{Verifies identification through heteroskedasticity or non-normality of 
of structural shocks}
\usage{
\method{verify_identification}{PosteriorBSVARMSH}(posterior)
}
\arguments{
\item{posterior}{the estimation outcome obtained using \code{\link{estimate}} function}
}
\value{
An object of class \code{SDDRid*} that is a list with components:

\code{logSDDR} a vector with values of the logarithm of the Bayes factors

\code{logSDDR_se} a vector with numerical standard errors of the logarithm of
the Bayes factors reported in output element \code{logSDDR} that are computed 
based on 30 random sub-samples of the log-ordinates of the marginal posterior 
and prior distributions.
}
\description{
Computes the logarithm of Bayes factor for the homoskedasticity hypothesis 
for each of the structural shocks via Savage-Dickey Density Ration (SDDR).
The hypothesis of homoskedasticity is represented by restriction:
\deqn{H_0: \sigma^2_{n.1} = ... = \sigma^2_{n.M} = 1}
The logarithm of Bayes factor for this hypothesis can be computed using the SDDR 
as the difference of logarithms of the marginal posterior distribution ordinate 
at the restriction less the marginal prior distribution ordinate at the same point:
\deqn{log p(H_0 | data) - log p(H_0)}
Therefore, a negative value of the difference is the evidence against 
homoskedasticity of the structural shock. The estimation of both elements of 
the difference requires numerical integration.
}
\examples{
# simple workflow
############################################################
# specify the model
specification  = specify_bsvar_msh$new(us_fiscal_lsuw, p = 1, M = 2)

# estimate the model
posterior      = estimate(specification, 10)

# verify heteroskedasticity
sddr           = verify_identification(posterior)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_msh$new(p = 1, M = 2) |>
  estimate(S = 10) |> 
  verify_identification() -> sddr
  
}
\references{
Lütkepohl, H., and Woźniak, T., (2020) Bayesian Inference for Structural Vector Autoregressions Identified by Markov-Switching Heteroskedasticity. \emph{Journal of Economic Dynamics and Control} \bold{113}, 103862, \doi{10.1016/j.jedc.2020.103862}.

Lütkepohl, H., Shang, F., Uzeda, L., and Woźniak, T. (2024) Partial Identification of Heteroskedastic Structural VARs: Theory and Bayesian Inference. \emph{University of Melbourne Working Paper}, 1--57, \doi{10.48550/arXiv.2404.11057}.
}
\seealso{
\code{\link{specify_bsvar_msh}}, \code{\link{estimate}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
