% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_normality.PosteriorBSVARHMSH}
\alias{verify_normality.PosteriorBSVARHMSH}
\title{Verifies normality of structural shocks equation by equation}
\usage{
\method{verify_normality}{PosteriorBSVARHMSH}(posterior)
}
\arguments{
\item{posterior}{the \code{posterior} element of the list from the estimation outcome}
}
\value{
An object of class \code{SDDRnormality} that is a list of three components:

\code{logSDDR} an \code{N}-vector with values of the logarithm of the Bayes factors for 
the normality hypothesis for each of the shocks

\code{log_SDDR_se} an \code{N}-vector with estimation standard errors of the logarithm of 
the Bayes factors reported in output element \code{logSDDR} that are computed based on 30 random 
sub-samples of the log-ordinates of the marginal posterior and prior distributions.
}
\description{
Computes the logarithm of Bayes factor for the normality hypothesis 
for each of the structural shocks via Savage-Dickey Density Ration (SDDR).
The hypothesis of normality, \eqn{H_0}, is represented by restriction that the
equation-specific degrees of freedom parameter is equal to infinity, 
\eqn{\nu_n\rightarrow\infty}.
The logarithm of Bayes factor for this hypothesis can be computed using the SDDR 
as the difference of logarithms of the marginal posterior distribution ordinate at the restriction 
less the marginal prior distribution ordinate at the same point:
\deqn{log p(H_0 | data) - log p(H_0)}
Therefore, a negative value of the difference is the evidence against 
normality of the structural shock. The estimation of th first element relies on 
kernel density estimation of the marginal posterior density, whereas the second
element is equal to the log of value 1.
}
\examples{
# simple workflow
############################################################
# specify the model 
specification  = specify_bsvar_msh$new(us_fiscal_lsuw, M = 2, distribution = "t")

# estimate the model
posterior      = estimate(specification, 10)

# verify heteroskedasticity
sddr           = verify_normality(posterior)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_msh$new(M = 2, distribution = "t") |>
  estimate(S = 10) |> 
  verify_normality() -> sddr
  
}
\references{
Lütkepohl, H., and Woźniak, T., (2020) Bayesian Inference for Structural Vector Autoregressions Identified by Markov-Switching Heteroskedasticity. \emph{Journal of Economic Dynamics and Control} \bold{113}, 103862, \doi{10.1016/j.jedc.2020.103862}.

Lütkepohl, H., Shang, F., Uzeda, L., and Woźniak, T. (2024) Partial Identification of Heteroskedastic Structural VARs: Theory and Bayesian Inference. \emph{University of Melbourne Working Paper}, 1--57, \doi{10.48550/arXiv.2404.11057}.
}
\seealso{
\code{\link{specify_bsvar_hmsh}}, \code{\link{estimate}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
