% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsvars-package.R
\docType{package}
\name{bsvars-package}
\alias{bsvars-package}
\alias{bsvars}
\title{Bayesian Estimation of Structural Vector Autoregressive Models}
\description{
Provides fast and efficient procedures for Bayesian analysis of 
Structural Vector Autoregressions. This package estimates a wide range of 
models, including homo-, heteroskedastic and non-normal specifications. 
Structural models can be identified by adjustable exclusion restrictions, 
time-varying volatility, or non-normality, and include exclusion restrictions 
on autoregressive parameters.
They all include a flexible three-level equation-specific local-global 
hierarchical prior distribution for the estimated level of shrinkage for 
autoregressive and structural parameters. Additionally, the package facilitates 
predictive and structural analyses such as impulse responses, forecast error 
variance and historical decompositions, forecasting, verification of 
heteroskedasticity and hypotheses on autoregressive parameters, and analyses 
of structural shocks, volatilities, and fitted values. Beautiful plots, 
informative summary functions, and extensive documentation including the 
vignette by Woźniak (2024) <doi:10.48550/arXiv.2410.15090> complement all this. 
The implemented techniques align closely with those presented in 
Lütkepohl, Shang, Uzeda, & Woźniak (2025) <doi:10.1016/j.jeconom.2025.106107>, 
Lütkepohl & Woźniak (2020) <doi:10.1016/j.jedc.2020.103862>, 
Song & Woźniak (2021) <doi:10.1093/acrefore/9780190625979.013.174>, and 
Woźniak & Droumaguet (2015) <doi:10.13140/RG.2.2.19492.55687>. The 'bsvars' 
package is aligned regarding objects, workflows, and code structure with the 
R package 'bsvarSIGNs' by Wang & Woźniak (2024) 
<doi:10.32614/CRAN.package.bsvarSIGNs>, and they constitute an integrated 
toolset.
}
\details{
\strong{Models.} All the SVAR models in this package are specified by two equations, including 
the reduced form equation:
\deqn{Y = AX + E}
where \eqn{Y} is an \code{NxT} matrix of dependent variables, 
\eqn{X} is a \code{KxT} matrix of explanatory variables, 
\eqn{E} is an \code{NxT} matrix of reduced form error terms, 
and \eqn{A} is an \code{NxK} matrix of autoregressive slope coefficients and 
parameters on deterministic terms in \eqn{X}.

The structural equation is given by:
\deqn{BE = U}
where \eqn{U} is an \code{NxT} matrix of structural form error terms, and
\eqn{B} is an \code{NxN} matrix of contemporaneous relationships.

Finally, all of the models share assumptions regarding the structural
shocks \code{U}, namely, temporal and contemporaneous independence. They imply 
zero correlations and autocorrelations. 

The various SVAR models estimated differ by the specification of structural shocks
variances. The different models include:
\itemize{
  \item homoskedastic model with unit variances
  \item heteroskedastic model with non-centred Stochastic Volatility process for variances
  \item heteroskedastic model with centred Stochastic Volatility process for variances
  \item heteroskedastic model with stationary Markov switching in the variances
  \item heteroskedastic model with sparse Markov switching in the variances where the number of heteroskedastic components is estimated
  \item heteroskedastic model with stationary heterogeneous Markov switching in the variances, where each shock volatility has its own Markov process
  \item heteroskedastic model with sparse heterogeneous Markov switching in the variances where the number of heteroskedastic components is estimated
  \item heteroskedastic model with exogenous heteroskedastic regime changes in the variances
  \item a model with Student-t distributed structural shocks with estimated equation-specific degrees-of-freedom parameter
  \item non-normal model with a finite mixture of normal components and component-specific variances
  \item non-normal model with a sparse mixture of normal components and component-specific variances where the number of heteroskedastic components is estimated
}
The structural shocks can be either normally or Student-t distributed, where in 
the latter case the shock-specific degrees of freedom parameters are estimated.

\strong{Prior distributions.} All the models feature a Minnesota prior for autoregressive 
parameters in matrix \eqn{A} and a generalised-normal distribution for the structural 
matrix \eqn{B}. Both of these distributions feature a 3-level equation-specific
local-global hierarchical prior that make the shrinkage estimation flexible improving
the model fit and its forecasting performance.

\strong{Estimation algorithm.} The models are estimated using frontier numerical methods
making the Gibbs sampler fast and efficient. The estimation follows closely 
Lütkepohl, Shang, Uzeda, & Woźniak (2025). The sampler of the structural matrix 
follows Waggoner & Zha (2003), whereas that 
for autoregressive parameters follows Chan, Koop, Yu (2022). 
The specification of Markov switching heteroskedasticity is inspired by 
Song & Woźniak (2021), and that of 
Stochastic Volatility model by Kastner & Frühwirth-Schnatter (2014).
The identification problems are considered in Lütkepohl, Shang, Uzeda, & Woźniak (2025)
and Lütkepohl & Woźniak (2020).

\strong{Identification verification.} The structural shocks can be identified 
through heteroskedasticity or non-normality following Lütkepohl, Shang, Uzeda, & Woźniak (2025)
and Lütkepohl & Woźniak (2020). The package provides functions to verify both,
homoskedasticity and normality of the structural shocks, which facilitates 
making probabilistic statements regarding the identification. Additionally,
the package makes it possible to verify linear restrictions on autoregressive 
parameters.
}
\note{
This package is currently in active development. Your comments,
suggestions and requests are warmly welcome!
}
\examples{
spec  = specify_bsvar_sv$new(         # specify the model
          us_fiscal_lsuw, 
          exogenous = us_fiscal_ex
        )
burn  = estimate(spec, 5)             # run the burn-in
post  = estimate(burn, 5)             # estimate the model
irf   = compute_impulse_responses(    # compute impulse responses
          post, 
          horizon = 2
         )

# compute forecast error variance decomposition one year ahead
fevd  = compute_variance_decompositions(post, horizon = 4)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_sv$new(exogenous = us_fiscal_ex) |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_variance_decompositions(horizon = 4) -> fevds

# conditional forecasting using a model with exogenous variables
############################################################
us_fiscal_lsuw |>
  specify_bsvar_sv$new(exogenous = us_fiscal_ex) |>
  estimate(S = 5) |> 
  estimate(S = 5) -> post
  
 post |> forecast(
    horizon = 8,
    exogenous_forecast = us_fiscal_ex_forecasts,
    conditional_forecast = us_fiscal_cond_forecasts
  ) -> pred
  
  pred |> summary()
  pred |> plot(probability = 0.68)
  
# estimation of a model with exogeneity restrictions on the  autoregressive matrix
#############################################################
A = matrix(TRUE, 3, 7)
A[1,3] = A[1,6] = FALSE
us_fiscal_lsuw |>
  specify_bsvar_sv$new(p = 2, A = A) |>
  estimate(S = 5) |> 
  estimate(S = 5) -> post
post |> summary()

}
\references{
Chan, J.C.C., Koop, G, and Yu, X. (2024) Large Order-Invariant Bayesian VARs with Stochastic Volatility. 
\emph{Journal of Business & Economic Statistics}, \bold{42}, \doi{10.1080/07350015.2023.2252039}.

Kastner, G. and Frühwirth-Schnatter, S. (2014) Ancillarity-Sufficiency Interweaving Strategy (ASIS) for Boosting MCMC 
Estimation of Stochastic Volatility Models. \emph{Computational Statistics & Data Analysis}, \bold{76}, 408--423, 
\doi{10.1016/j.csda.2013.01.002}.

Lütkepohl, H., Shang, F., Uzeda, L., and Woźniak, T. (2025) 
Partial identification of structural vector autoregressions with non-centred stochastic volatility. 
\emph{Journal of Econometrics} \bold{256}, 106107, \doi{10.1016/j.jeconom.2025.106107}.

Lütkepohl, H., and Woźniak, T., (2020) Bayesian Inference for Structural Vector Autoregressions Identified by Markov-Switching Heteroskedasticity. 
\emph{Journal of Economic Dynamics and Control} \bold{113}, 103862, \doi{10.1016/j.jedc.2020.103862}.

Song, Y., and Woźniak, T. (2021) Markov Switching Heteroskedasticity in Time Series Analysis. 
In: \emph{Oxford Research Encyclopedia of Economics and Finance}. Oxford University Press, \doi{10.1093/acrefore/9780190625979.013.174}.

Waggoner, D.F., and Zha, T., (2003) A Gibbs sampler for structural vector autoregressions. 
\emph{Journal of Economic Dynamics and Control}, \bold{28}, 349--366, \doi{10.1016/S0165-1889(02)00168-9}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bsvars.org/bsvars/}
  \item Report bugs at \url{https://github.com/bsvars/bsvars/issues}
}

}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
\keyword{models}
\keyword{package}
\keyword{ts}
