% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsvars-package.R
\docType{package}
\name{bsvars-package}
\alias{bsvars-package}
\alias{bsvars}
\title{Bayesian Estimation of Structural Vector Autoregressive Models}
\description{
Efficient algorithms for Bayesian estimation of 
Structural Vector Autoregressive (SVAR) models via Markov chain Monte Carlo methods. 
A wide range of SVAR models is considered, including homo- and heteroskedastic specifications 
and those with non-normal structural shocks.
}
\details{
All the SVAR models in this package are specified by two equations, including 
the reduced form equation:
\deqn{Y = AX + E}
where \eqn{Y} is an \code{NxT} matrix of dependent variables, 
\eqn{X} is a \code{KxT} matrix of explanatory variables, 
\eqn{E} is an \code{NxT} matrix of reduced form error terms, 
and \eqn{A} is an \code{NxK} matrix of autoregressive slope coefficients and 
parameters on deterministic terms in \eqn{X}.

The structural equation is given by:
\deqn{BE = U}
where \eqn{U} is an \code{NxT} matrix of structural form error terms, and
\eqn{B} is an \code{NxN} matrix of contemporaneous relationships.

Finally, all of the models share the following assumptions regarding the structural
shocks \code{U}, namely, joint conditional normality given the past observations collected
in matrix \code{X}, and temporal and contemporaneous independence. The latter implies 
zero correlations and autocorrelations. 

The various SVAR models estimated differ by the specification of structural shocks
variances. The different models include:
\itemize{
  \item homoskedastic model with unit variances
  \item heteroskedastic model with stationary Markov switching in the variances
  \item heteroskedastic model with Stochastic Volatility process for variances
  \item non-normal model with a finite mixture of normal components and component-specific variances
  \item heteroskedastic model with sparse Markov switching in the variances where the number of heteroskedastic components is estimated
  \item non-normal model with a sparse mixture of normal components and component-specific variances where the number of heteroskedastic components is estimated
}
}
\note{
This package is currently in active development. Your comments,
suggestions and requests are warmly welcome!
}
\examples{
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_sv$new(us_fiscal_lsuw, p = 4)
set.seed(123)

# run the burn-in
burn_in        = estimate_bsvar_sv(10, specification)

# estimate the model
posterior      = estimate_bsvar_sv(50, burn_in$get_last_draw())

# normalise the posterior
BB            = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat         = diag(sign(diag(BB))) \%*\% BB                # set positive diagonal elements
bsvars::normalise_posterior(posterior, B_hat)              # draws in posterior are normalised
}
\references{
Woźniak, T., and Droumaguet, M., (2022) Bayesian Assessment of Identifying Restrictions for Heteroskedastic Structural VARs.
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
\keyword{models}
\keyword{package}
\keyword{ts}
