\name{bsvars-package}
\alias{bsvars-package}
\alias{bsvars}
\docType{package}
\title{
\packageTitle{bsvars}
}
\description{
\packageDescription{bsvars}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bsvars}
\packageIndices{bsvars}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bsvars}

Maintainer: \packageMaintainer{bsvars}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
