% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_conditional_sd.R
\name{compute_conditional_sd}
\alias{compute_conditional_sd}
\title{Computes posterior draws of structural shock conditional standard deviations}
\usage{
compute_conditional_sd(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome obtained by running the 
\code{estimate} function. The interpretation depends on the normalisation of 
the shocks using function \code{normalise_posterior()}. Verify if the default 
settings are appropriate.}
}
\value{
An object of class \code{PosteriorSigma}, that is, an \code{NxTxS} 
array with attribute \code{PosteriorSigma} containing \code{S} draws of the 
structural shock conditional standard deviations.
}
\description{
Each of the draws from the posterior estimation of models is 
transformed into a draw from the posterior distribution of the structural 
shock conditional standard deviations.
}
\examples{
specification  = specify_bsvar$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)
sigma          = compute_conditional_sd(posterior)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_conditional_sd() -> csd

}
\seealso{
\code{\link{estimate}}, \code{\link{normalise_posterior}}, \code{\link{summary}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
