% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_historical_decompositions.R
\name{compute_historical_decompositions}
\alias{compute_historical_decompositions}
\title{Computes posterior draws of historical decompositions}
\usage{
compute_historical_decompositions(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of either of the classes: 
PosteriorBSVAR, PosteriorBSVAR-MSH, PosteriorBSVAR-MIX, or PosteriorBSVAR-SV
obtained by running one of the \code{estimate_bsvar_*} functions. The draws must be normalised
using function \code{normalise_posterior()} for the historical decompositions to be interpretable.}
}
\value{
An object of class PosteriorHD, that is, an \code{NxNxTxS} array with attribute PosteriorHD 
containing \code{S} draws of the historical decompositions.
}
\description{
Each of the draws from the posterior estimation of a model is transformed into
a draw from the posterior distribution of the historical decompositions.
}
\examples{
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)

# run the burn-in
burn_in        = estimate_bsvar(specification, 10)

# estimate the model
posterior      = estimate_bsvar(burn_in$get_last_draw(), 50)

# compute historical decompositions
hd            = compute_historical_decompositions(posterior)

}
\references{
Kilian, L., & Lütkepohl, H. (2017). Structural VAR Tools, Chapter 4, In: Structural vector autoregressive analysis. Cambridge University Press.
}
\seealso{
\code{\link{estimate_bsvar}}, \code{\link{estimate_bsvar_msh}}, \code{\link{estimate_bsvar_sv}}, \code{\link{estimate_bsvar_mix}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
