% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_impulse_responses.R
\name{compute_impulse_responses}
\alias{compute_impulse_responses}
\title{Computes posterior draws of impulse responses}
\usage{
compute_impulse_responses(posterior, horizon)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of either of the classes: 
PosteriorBSVAR, PosteriorBSVAR-MSH, PosteriorBSVAR-MIX, or PosteriorBSVAR-SV
obtained by running one of the \code{estimate_bsvar_*} functions. The draws must be normalised
using function \code{normalise_posterior()} for the impulse responses to be interpretable.}

\item{horizon}{a positive integer number denoting the forecast horizon for the impulse responses computations.}
}
\value{
An object of class PosteriorIR, that is, an \code{NxNx(horizon+1)xS} array with attribute PosteriorIR 
containing \code{S} draws of the impulse responses.
}
\description{
Each of the draws from the posterior estimation of a model is transformed into
a draw from the posterior distribution of the impulse responses.
}
\examples{
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)

# run the burn-in
burn_in        = estimate_bsvar(specification, 10)

# estimate the model
posterior      = estimate_bsvar(burn_in$get_last_draw(), 50)

# compute impulse responses 5 years ahead
irf           = compute_impulse_responses(posterior, horizon = 20)

}
\references{
Kilian, L., & Lütkepohl, H. (2017). Structural VAR Tools, Chapter 4, In: Structural vector autoregressive analysis. Cambridge University Press.
}
\seealso{
\code{\link{estimate_bsvar}}, \code{\link{estimate_bsvar_msh}}, \code{\link{estimate_bsvar_sv}}, \code{\link{estimate_bsvar_mix}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
