% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_structural_shocks.R
\name{compute_structural_shocks}
\alias{compute_structural_shocks}
\title{Computes posterior draws of structural shocks}
\usage{
compute_structural_shocks(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of either of the classes: 
PosteriorBSVAR, PosteriorBSVAR-MSH, PosteriorBSVAR-MIX, or PosteriorBSVAR-SV
obtained by running one of the \code{estimate_bsvar_*} functions. The draws must be normalised
using function \code{normalise_posterior()} for the structural shocks to be interpretable.}
}
\value{
An object of class PosteriorShocks, that is, an \code{NxTxS} array with attribute PosteriorShocks 
containing \code{S} draws of the structural shocks.
}
\description{
Each of the draws from the posterior estimation of a model is transformed into
a draw from the posterior distribution of the structural shocks.
}
\examples{
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)

# run the burn-in
burn_in        = estimate_bsvar(specification, 10)

# estimate the model
posterior      = estimate_bsvar(burn_in$get_last_draw(), 50)

# compute impulse responses 5 years ahead
shocks         = compute_structural_shocks(posterior)

}
\seealso{
\code{\link{estimate_bsvar}}, \code{\link{estimate_bsvar_msh}}, \code{\link{estimate_bsvar_sv}}, \code{\link{estimate_bsvar_mix}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
