% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variance_decompositions.R
\name{compute_variance_decompositions}
\alias{compute_variance_decompositions}
\title{Computes posterior draws of the forecast error variance decomposition}
\usage{
compute_variance_decompositions(posterior, horizon)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of one of the classes: 
PosteriorBSVAR, PosteriorBSVAR-MSH, PosteriorBSVAR-MIX, or PosteriorBSVAR-SV
obtained by running one of the \code{estimate_bsvar_*} functions - these draws must be normalised
using function \code{normalise_posterior()} for the forecast error variance decomposition to be interpretable.}

\item{horizon}{a positive integer number denoting the forecast horizon for the impulse responses computations.}
}
\value{
An object of class PosteriorFEVD, that is, an \code{NxNx(horizon+1)xS} array with attribute PosteriorFEVD 
containing \code{S} draws of the forecast error variance decomposition.
}
\description{
Each of the draws from the posterior estimation of a model
is transformed into a draw from the posterior distribution of the forecast error variance decomposition.
}
\examples{
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 4)

# run the burn-in
burn_in        = estimate_bsvar(10, specification)

# estimate the model
posterior      = estimate_bsvar(50, burn_in$get_last_draw())

# normalise the posterior
BB             = posterior$last_draw$starting_values$B      # get the last draw of B
B_hat          = diag(sign(diag(BB))) \%*\% BB                # set positive diagonal elements
normalise_posterior(posterior, B_hat)              # draws in posterior are normalised

# compute forecast error variance decomposition 5 years ahead
fevd           = compute_variance_decompositions(posterior, horizon = 20)

}
\references{
Kilian, L., & Lütkepohl, H. (2017). Structural VAR Tools, Chapter 4, In: Structural vector autoregressive analysis. Cambridge University Press.
}
\seealso{
\code{\link{compute_impulse_responses}}, \code{\link{estimate_bsvar}}, \code{\link{estimate_bsvar_msh}}, \code{\link{estimate_bsvar_sv}}, \code{\link{estimate_bsvar_mix}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
