% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.PosteriorBSVARSV}
\alias{forecast.PosteriorBSVARSV}
\title{Forecasting using Bayesian Structural Vector Autoregression}
\usage{
\method{forecast}{PosteriorBSVARSV}(
  object,
  horizon = 1,
  exogenous_forecast = NULL,
  conditional_forecast = NULL,
  ...
)
}
\arguments{
\item{object}{posterior estimation outcome - an object of class 
\code{PosteriorBSVARSV} obtained by running the \code{estimate} function.}

\item{horizon}{a positive integer, specifying the forecasting horizon.}

\item{exogenous_forecast}{a matrix of dimension \code{horizon x d} containing 
forecasted values of the exogenous variables.}

\item{conditional_forecast}{a \code{horizon x N} matrix with forecasted values 
for selected variables. It should only contain \code{numeric} or \code{NA} 
values. The entries with \code{NA} values correspond to the values that are 
forecasted conditionally on the realisations provided as \code{numeric} values.}

\item{...}{not used}
}
\value{
A list of class \code{Forecasts} containing the
draws from the predictive density and for heteroskedastic models the draws 
from the predictive density of structural shocks conditional standard 
deviations and data. The output elements include:

\describe{
 \item{forecasts}{an \code{NxTxS} array with the draws from predictive density}
 \item{Y}{an \eqn{NxT} matrix with the data on dependent variables}
 \item{forecast_mean}{an \code{NxTxS} array with the mean of the predictive density}
 \item{forecast_covariance}{an \code{NxTxS} array with the covariance of the predictive density}
}
}
\description{
Samples from the joint predictive density of all of the dependent 
variables for models at forecast horizons from 1 to \code{horizon} specified as 
an argument of the function.
}
\examples{
specification  = specify_bsvar_sv$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)
predictive     = forecast(posterior, 2)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_sv$new() |>
  estimate(S = 5) |>
  estimate(S = 5) |>  
  forecast(horizon = 2) -> predictive
  
# conditional forecasting using a model with exogenous variables
############################################################
specification  = specify_bsvar_sv$new(us_fiscal_lsuw, exogenous = us_fiscal_ex)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# forecast 2 years ahead
predictive     = forecast(
                    posterior, 
                    horizon = 8,
                    exogenous_forecast = us_fiscal_ex_forecasts,
                    conditional_forecast = us_fiscal_cond_forecasts
                  )
summary(predictive)

# workflow with the pipe |>
############################################################
us_fiscal_lsuw |>
  specify_bsvar_sv$new(exogenous = us_fiscal_ex) |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  forecast(
    horizon = 8,
    exogenous_forecast = us_fiscal_ex_forecasts,
    conditional_forecast = us_fiscal_cond_forecasts
  ) |> plot()

}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
