% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorFitted}
\alias{summary.PosteriorFitted}
\title{Provides posterior summary of variables' fitted values}
\usage{
\method{summary}{PosteriorFitted}(object, ...)
}
\arguments{
\item{object}{an object of class PosteriorFitted obtained using the
\code{compute_fitted_values()} function containing draws the predictive 
density of the sample data.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 
5 and 95 percentiles of the fitted values for each of the shocks and periods.
}
\description{
Provides posterior summary of the fitted values including their 
mean, standard deviations, as well as 5 and 95 percentiles.
}
\examples{
specification  = specify_bsvar$new(us_fiscal_lsuw)
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# compute fitted values
fitted         = compute_fitted_values(posterior)
fitted_summary = summary(fitted)
fitted_summary$variable1[,1] # access posterior mean of variable1

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_fitted_values() |>
  summary() -> fitted_summary
fitted_summary$variable1[,1] # access posterior mean of variable1

}
\seealso{
\code{\link{compute_fitted_values}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
