% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorHD}
\alias{summary.PosteriorHD}
\title{Provides posterior summary of historical decompositions}
\usage{
\method{summary}{PosteriorHD}(object, ...)
}
\arguments{
\item{object}{an object of class PosteriorHD obtained using the
\code{compute_historical_decompositions()} function containing posterior draws
 of historical decompositions.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior means of historical decompositions for
each of the variables.
}
\description{
Provides posterior means of the historical decompositions variable
by variable.
}
\examples{
specification  = specify_bsvar$new(diff(us_fiscal_lsuw))
burn_in        = estimate(specification, 5)
posterior      = estimate(burn_in, 5)

# compute historical decompositions
hds            = compute_historical_decompositions(posterior)
hds_summary    = summary(hds)
head(hds_summary$gdp) # browse the contributions

# workflow with the pipe |>
############################################################
diff(us_fiscal_lsuw) |>
  specify_bsvar$new() |>
  estimate(S = 5) |> 
  estimate(S = 5) |> 
  compute_historical_decompositions() |>
  summary() -> hds_summary
hds_summary$gdp # browse the contributions

}
\seealso{
\code{\link{compute_historical_decompositions}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
