% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PosteriorIR}
\alias{summary.PosteriorIR}
\title{Provides posterior summary of variables' fitted values}
\usage{
\method{summary}{PosteriorIR}(object, ...)
}
\arguments{
\item{object}{an object of class PosteriorIR obtained using the
\code{compute_impulse_responses()} function containing draws the predictive 
density of the sample data.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A list reporting the posterior mean, standard deviations, as well as 
5 and 95 percentiles of the impulse responses of each variable to each of the 
shocks at all horizons.
}
\description{
Provides posterior summary of the fitted values including their 
mean, standard deviations, as well as 5 and 95 percentiles.
}
\examples{
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
set.seed(123)
specification  = specify_bsvar$new(us_fiscal_lsuw, p = 1)

# run the burn-in
burn_in        = estimate(specification, 10)

# estimate the model
posterior      = estimate(burn_in$get_last_draw(), 20, , thin = 1)

# compute impulse responses
irf            = compute_impulse_responses(posterior, horizon = 8)
irf_summary    = summary(irf)

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar$new(p = 1) |>
  estimate(S = 10) |> 
  estimate(S = 20, thin = 1) |> 
  compute_impulse_responses(horizon = 8) |>
  summary() -> irf_summary

}
\seealso{
\code{\link{compute_impulse_responses}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
