% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_volatility.PosteriorBSVARSV}
\alias{verify_volatility.PosteriorBSVARSV}
\title{Verifies heteroskedasticity of structural shocks equation by equation}
\usage{
\method{verify_volatility}{PosteriorBSVARSV}(posterior)
}
\arguments{
\item{posterior}{the \code{posterior} element of the list from the estimation outcome}
}
\value{
An object of class SDDR that is a list of three components:

\code{logSDDR} an \code{N}-vector with values of the logarithm of the Bayes factors for 
the homoskedasticity hypothesis for each of the shocks

\code{log_SDDR_se} an \code{N}-vector with estimation standard errors of the logarithm of 
the Bayes factors reported in output element \code{logSDDR} that are computed based on 30 random 
sub-samples of the log-ordinates of the marginal posterior and prior distributions.

\code{components} a list of three components for the computation of the Bayes factor
\describe{
  \item{log_denominator}{an \code{N}-vector with values of the logarithm of the Bayes factor denominators}
  \item{log_numerator}{an \code{N}-vector with values of the logarithm of the Bayes factor numerators}
  \item{log_numerator_s}{an \code{NxS} matrix of the log-full conditional posterior density ordinates computed to estimate the numerator}
  \item{se_components}{an \code{Nx30} matrix containing the log-Bayes factors on the basis of which the standard errors are computed}
}
}
\description{
Computes the logarithm of Bayes factor for the homoskedasticity hypothesis 
for each of the structural shocks via Savage-Dickey Density Ration (SDDR).
The hypothesis of homoskedasticity is represented by restriction:
\deqn{H_0: \omega_n = 0}
The logarithm of Bayes factor for this hypothesis can be computed using the SDDR 
as the difference of logarithms of the marginal posterior distribution ordinate at the restriction 
less the marginal prior distribution ordinate at the same point:
\deqn{log p(\omega_n = 0 | data) - log p(\omega_n = 0)}
Therefore, a negative value of the difference is the evidence against 
homoskedasticity of the structural shock. The estimation of both elements of the difference requires 
numerical integration.
}
\examples{
# simple workflow
############################################################
# upload data
data(us_fiscal_lsuw)

# specify the model and set seed
specification  = specify_bsvar_sv$new(us_fiscal_lsuw, p = 1)
set.seed(123)

# run the burn-in
posterior      = estimate(specification, 60, thin = 1)

# verify heteroskedasticity
sddr           = verify_volatility(posterior)

# workflow with the pipe |>
############################################################
set.seed(123)
us_fiscal_lsuw |>
  specify_bsvar_sv$new(p = 1) |>
  estimate(S = 60, thin = 1) |> 
  verify_volatility() -> sddr
  
}
\references{
Sampling from the generalised-normal full conditional posterior distribution of matrix \eqn{B} is implemented using the Gibbs sampler by:

Waggoner, D.F., and Zha, T., (2003) A Gibbs sampler for structural vector autoregressions. \emph{Journal of Economic Dynamics and Control}, \bold{28}, 349--366, \doi{https://doi.org/10.1016/S0165-1889(02)00168-9}.

Sampling from the multivariate normal full conditional posterior distribution of each of the \eqn{A} matrix row is implemented using the sampler by:

Chan, J.C.C., Koop, G, and Yu, X. (2021) Large Order-Invariant Bayesian VARs with Stochastic Volatility.
}
\seealso{
\code{\link{specify_bsvar_sv}}, \code{\link{estimate}}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
